document-processor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
sentence-transformers==2.2.2
chromadb==0.4.15
huggingface-hub==0.16.4
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
numpy==1.24.4
